requirements.txt
langchain==0.3.27
langchain-core~=0.3.75
langchain-qwq==0.2.1
langgraph==0.6.6
ipython==9.5.0
python-dotenv==1.1.1
tavily-python==0.7.11
langchain-tavily==0.2.11
